updated account fields data types for dieselgit add .!

+++ src/models/AccountModel.rs
@@ -2,26 +2,26 @@ use std::str::FromStr;
 
 #[derive(Debug, Queryable)]
 pub struct Account {
-    pub no: u32,
+    pub no: i32,
     pub name: String,
-    pub deposit: i128,
+    pub deposit: f64,
 }
 
 impl Account {
-    pub fn new(no: u32, name: String, deposit: i128) -> Account {
+    pub fn new(no: i32, name: String, deposit: f64) -> Account {
         Account { no, name, deposit }
     }
 
-    pub fn Deposit(&mut self, amount: i128) -> &'static str {
+    pub fn Deposit(&mut self, amount: f64) -> &'static str {
         self.deposit = self.deposit + amount;
         return "Deposit successfully";
     }
 
-    pub fn CanWithdraw(&mut self, amount: i128) -> bool {
+    pub fn CanWithdraw(&mut self, amount: f64) -> bool {
         return self.deposit >= amount;
     }
 
-    pub fn Withdraw(&mut self, amount: i128) -> Result<&'static str, &'static str> {
+    pub fn Withdraw(&mut self, amount: f64) -> Result<&'static str, &'static str> {
         return match self.deposit < amount {
             true => Err("You don't have enough deposit to withdraw"),
             false => {
@@ -43,9 +43,9 @@ impl Account {
 
     pub fn from_str(account_details: String) -> Account {
         let split_into: Vec<&str> = account_details.split(',').collect();
-        let no: u32 = FromStr::from_str(split_into[0]).unwrap();
+        let no: i32 = FromStr::from_str(split_into[0]).unwrap();
         let name: String = split_into[1].to_string();
-        let deposit: i128 = <i128 as FromStr>::from_str(split_into[2]).unwrap();
+        let deposit: f64 = <f64 as FromStr>::from_str(split_into[2]).unwrap();
         Account {
             no,
             name,
@@ -60,40 +60,40 @@ mod tests {
     use super::Account;
     #[test]
     fn should_deposit() {
-        let mut account: Account = Account::new(1, "Jack".to_string(), 100);
-        account.Deposit(500);
-        assert_eq!(account.deposit, 600);
+        let mut account: Account = Account::new(1, "Jack".to_string(), 100.0);
+        account.Deposit(500.0);
+        assert_eq!(account.deposit, 600.0);
 
-        account.Deposit(700);
-        assert_eq!(account.deposit, 1300);
+        account.Deposit(700.0);
+        assert_eq!(account.deposit, 1300.0);
     }
 
     #[test]
     fn should_withdraw() {
-        let mut account: Account = Account::new(1, "Jack".to_string(), 100);
-        assert_eq!(account.Withdraw(100), Ok("Withdraw successfully"));
+        let mut account: Account = Account::new(1, "Jack".to_string(), 100.0);
+        assert_eq!(account.Withdraw(100.0), Ok("Withdraw successfully"));
     }
 
     #[test]
     fn should_not_withdraw() {
-        let mut account: Account = Account::new(1, "Jack".to_string(), 100);
+        let mut account: Account = Account::new(1, "Jack".to_string(), 100.0);
         assert_eq!(
-            account.Withdraw(101),
+            account.Withdraw(101.0),
             Err("You don't have enough deposit to withdraw")
         );
     }
 
     #[test]
     fn should_determine_withdrawl() {
-        let mut account: Account = Account::new(1, "Jack".to_string(), 100);
-        assert_eq!(account.CanWithdraw(50), true);
-        assert_eq!(account.CanWithdraw(100), true);
-        assert_eq!(account.CanWithdraw(101), false);
+        let mut account: Account = Account::new(1, "Jack".to_string(), 100.0);
+        assert_eq!(account.CanWithdraw(50.0), true);
+        assert_eq!(account.CanWithdraw(100.0), true);
+        assert_eq!(account.CanWithdraw(101.0), false);
     }
 
     #[test]
     fn should_convert_csv_string() {
-        let mut account: Account = Account::new(1, "Jack".to_string(), 100);
+        let mut account: Account = Account::new(1, "Jack".to_string(), 100.0);
         let string: String = account.Stringify();
         assert_eq!(string, String::from("\n1,Jack,100"));
     }
