+++ src/models/TransferModel.rs
@@ -1,5 +1,5 @@
 pub struct Transfer {
-    pub from: u32,
-    pub to: u32,
-    pub amount: i128,
+    pub from: i32,
+    pub to: i32,
+    pub amount: f64,
 }
